# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth-rest-api/user/auth/refreshToken": {
            "post": {
                "description": "Refresh Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users V1"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "P-DeviceType Header",
                        "name": "P-DeviceType",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization Header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apihelpers.APIRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.AuthToken"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.APIRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.APIRes"
                        }
                    }
                }
            }
        },
        "/api/auth-rest-api/user/auth/revokeToken": {
            "post": {
                "description": "Revoke Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users V1"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "P-DeviceType Header",
                        "name": "P-DeviceType",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization Header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.APIRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.APIRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.APIRes"
                        }
                    }
                }
            }
        },
        "/api/auth-rest-api/user/signIn": {
            "post": {
                "description": "enter email and password for signin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users V1"
                ],
                "parameters": [
                    {
                        "description": "signin",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LogInReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apihelpers.APIRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.LogInRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.APIRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.APIRes"
                        }
                    }
                }
            }
        },
        "/api/auth-rest-api/user/signUp": {
            "post": {
                "description": "enter email and password for signup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users V1"
                ],
                "parameters": [
                    {
                        "description": "Signup",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LogInReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apihelpers.APIRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.LogInRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.APIRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apihelpers.APIRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apihelpers.APIRes": {
            "type": "object",
            "properties": {
                "data": {},
                "errorcode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "models.AuthToken": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "models.LogInReq": {
            "type": "object",
            "properties": {
                "emailId": {
                    "type": "string",
                    "example": "abc@gmail.com"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.LogInRes": {
            "type": "object",
            "properties": {
                "authorization": {
                    "type": "string"
                }
            }
        }
    }
}